% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeaks.R
\docType{methods}
\name{findPeaks,MassSpectrum-method}
\alias{findPeaks,MassSpectrum-method}
\title{Find peaks in mass spectra}
\usage{
\S4method{findPeaks}{MassSpectrum}(object, SNR = 6, ...)
}
\arguments{
\item{object}{A mass spectrum}

\item{SNR}{Signal-to-noise ratio}

\item{...}{Additional arguments}
}
\value{
A \code{MassSpectrum} object with peaks
}
\description{
Find peaks in mass spectra
}
