% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassSpectraList.R
\docType{class}
\name{MassSpectraList-class}
\alias{MassSpectraList-class}
\title{Mass spectra list objects}
\description{
\code{MassSpectraList} is a simple S4 class designed to
contain a list of \code{MassSpectrum} objects.
}
