% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassSpectrum.R
\docType{class}
\name{MassSpectrum-class}
\alias{MassSpectrum-class}
\alias{mz,MassSpectrum-method}
\alias{mz<-,MassSpectrum-method}
\alias{intensity,MassSpectrum-method}
\alias{peaks,MassSpectrum-method}
\alias{isCentroided,MassSpectrum-method}
\alias{tic,MassSpectrum-method}
\alias{length,MassSpectrum-method}
\alias{[,MassSpectrum,ANY,ANY,ANY-method}
\title{Mass spectrum objects}
\usage{
\S4method{mz}{MassSpectrum}(object, ...)

\S4method{mz}{MassSpectrum}(object) <- value

\S4method{intensity}{MassSpectrum}(object, ...)

\S4method{peaks}{MassSpectrum}(object, ...)

\S4method{isCentroided}{MassSpectrum}(object, ...)

\S4method{tic}{MassSpectrum}(object, ...)

\S4method{length}{MassSpectrum}(x)

\S4method{[}{MassSpectrum,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{object}{A \code{MassSpectrum} object}

\item{...}{Additional arguments}

\item{value}{A replacement value}
}
\description{
\code{MassSpectrum} is a simple S4 class designed to
encapsulate basic information about a mass spectrum.
}
\section{Methods (by generic)}{
\itemize{
\item \code{mz}: Get m/z values

\item \code{mz<-}: Set m/z values

\item \code{intensity}: Get intensities

\item \code{peaks}: Get indices of peaks

\item \code{isCentroided}: Get whether the spectrum is centroided

\item \code{tic}: Get the total ion count

\item \code{length}: Get the length of the spectrum

\item \code{[}: Subset a mass spectrum
}}

\section{Slots}{

\describe{
\item{\code{mz}}{The m/z-values.}

\item{\code{intensity}}{The intensities observed
at the associated m/z values.}

\item{\code{isCentroided}}{Whether the spectrum has
been centroided or not.}

\item{\code{peaks}}{The indices of any detected peaks}
}}

