% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassSpectrum.R
\name{MassSpectrum}
\alias{MassSpectrum}
\title{Create a \code{MassSpectrum} object.}
\usage{
MassSpectrum(mz, intensity, isCentroided = FALSE)
}
\arguments{
\item{mz}{A \code{numeric} vector of m/z values.}

\item{intensity}{A \code{numeric} vector of intensities.}

\item{isCentroided}{A \code{logical} value indicating whether
the spectrum has been centroided or not.}
}
\value{
object A \code{MassSpectrum} object.
}
\description{
Create a \code{MassSpectrum} object.
}
