% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSExample.R
\docType{package}
\name{MSExample-package}
\alias{MSExample-package}
\title{MSExample: Proteomics Experiment example package}
\description{
MSExample is an example R package that implements
a simple mass spectrometry pre-processing workflow for
the purpose of teaching Bioconductor package development.
}
