% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MassSpectrum.R
\docType{methods}
\name{plot,MassSpectrum,missing-method}
\alias{plot,MassSpectrum,missing-method}
\alias{plot,MassSpectrum,MassSpectrum-method}
\title{Plot a mass spectrum}
\usage{
\S4method{plot}{MassSpectrum,missing}(x, xlab = expression(italic(m/z)),
  ylab = expression(italic(Intensity)), type = if (isCentroided(x)) "h"
  else "l", col = "darkcyan", peaklabels = 10, add = FALSE, ...)

\S4method{plot}{MassSpectrum,MassSpectrum}(x, y,
  xlab = expression(italic(m/z)), ylab = expression(italic(Intensity)),
  type = if (isCentroided(x)) "h" else "l", col = c("darkcyan",
  "darkred"), peaklabels = 10, add = FALSE, ...)
}
\arguments{
\item{x}{A mass spectrum to plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{type}{The type of plot to draw}

\item{col}{The plot color}

\item{peaklabels}{Number of peaks to label}

\item{add}{Add to an existing plot?}

\item{...}{Arguments passed to plot function}

\item{y}{A second mass spectrum to plot}
}
\description{
Plots one or two mass spectra.
}
