% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeBaseline.R
\docType{methods}
\name{removeBaseline,MassSpectrum-method}
\alias{removeBaseline,MassSpectrum-method}
\alias{removeBaseline}
\title{Remove baseline from mass spectra}
\usage{
\S4method{removeBaseline}{MassSpectrum}(object, ...)
}
\arguments{
\item{object}{A mass spectrum}

\item{...}{Additional arguments}
}
\value{
A \code{MassSpectrum} object with baseline removed
}
\description{
Remove baseline from mass spectra
}
