% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSpectra.R
\name{simSpectra}
\alias{simSpectra}
\title{Simulate a mass spectrum}
\usage{
simSpectra(n = 1L, peaks = 50L, mz = rlnorm(peaks, 7, 0.3),
  intensity = rlnorm(peaks, 1, 0.9), from = 0.9 * min(mz), to = 1.1 *
  max(mz), by = 100, sdpeaks = sdpeakmult * log1p(intensity),
  sdpeakmult = 0.2, sdnoise = 0.1, sdmz = 10, resolution = 1000,
  fmax = 0.5, baseline = 0, decay = 10, units = c("ppm", "mz"),
  representation = c("profile", "centroid"), BPPARAM = bpparam())
}
\arguments{
\item{n}{The number of spectra to simulate}

\item{peaks}{The number of peaks to simulate}

\item{mz}{The m/z of the peaks}

\item{intensity}{The intensity of the peaks}

\item{from}{Lower end of mass range}

\item{to}{Upper end of mass range}

\item{by}{Size of m/z bins}

\item{sdpeaks}{Log-standard deviation of peak intensities}

\item{sdpeakmult}{Multiplier for peak variance}

\item{sdnoise}{Standard deviation of noise}

\item{sdmz}{Mass error}

\item{resolution}{Mass resolution as defined by 'm / dm'
where 'm' is the observed mass and 'dm'
is the width of the peak at a proportion
of its maximum height defined by \code{fmax}
(defaults to FWHM)}

\item{fmax}{The proportion of the maximum peak height used
by \code{resolution}}

\item{baseline}{Maximum intensity of hte baseline}

\item{decay}{Rate of decay of the baseline}

\item{units}{Units for the width of the m/z bins}

\item{representation}{Return a profile or centroided spectrum?}

\item{BPPARAM}{A parallel backend}
}
\value{
A \code{MassSpectrum} or \code{MassSpectraList}
}
\description{
Simulate a mass spectrum including peaks, noise,
a baseline, noise, mass drift, mass resolution, etc.
}
