% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothNoise.R
\docType{methods}
\name{smoothNoise,MassSpectrum-method}
\alias{smoothNoise,MassSpectrum-method}
\alias{smoothNoise}
\title{Smooth mass spectra}
\usage{
\S4method{smoothNoise}{MassSpectrum}(object, halfWindow = 2, ...)
}
\arguments{
\item{object}{A mass spectrum}

\item{halfWindow}{Smoothing window half-size}

\item{...}{Additional arguments}
}
\value{
A smoothed \code{MassSpectrum} object
}
\description{
Smooth mass spectra
}
